final cleanups
